extract PersistPostsService
