+++ gatekeeper-bot/src/main/kotlin/org/regenagcoop/discord/service/PersistedActivityService.kt
@@ -5,7 +5,6 @@ import org.regenagcoop.discord.Discord
 import org.regenagcoop.discord.model.Message
 import org.regenagcoop.discord.model.UserId
 import org.regenagcoop.model.ActiveMemberConfig
-import org.regenagcoop.model.PostHistory
 import java.time.LocalDate
 
 class PersistedActivityService(
@@ -57,45 +56,6 @@ class PersistedActivityService(
         }
 
         return persistedHistoryByDate
-    }
-
-    // TODO cleanup: extract to own parallel class of PersistReactionService
-    suspend fun persistPostHistoryForDay(date: LocalDate, usersWhoPostedOnDate: Set<UserId>) {
-        val usersStr = usersWhoPostedOnDate.sorted().joinToString()
-        val message = "Users who posted on $date: $usersStr"
-        discord.rooms.postMessage(message, activeMemberConfig.persistenceConfig.channel)
-    }
-
-    suspend fun persistMissingPostHistory(
-        today: LocalDate,
-        loadedPostHistory: PostHistory,
-        persistedDates: Set<LocalDate>
-    ) {
-        val startOfRelevantHistory = activeMemberConfig.computeEarliestScanDate(today)
-        val yesterday = today.minusDays(1)
-
-        // invert the map from UserId -> Dates to Date -> UserIds
-        val usersWhoPostedByDate = mutableMapOf<LocalDate, MutableSet<UserId>>()
-        loadedPostHistory.forEach { (userId, datesUserPosted) ->
-            datesUserPosted.forEach { dateUserPosted ->
-                if (dateUserPosted !in usersWhoPostedByDate) {
-                    usersWhoPostedByDate[dateUserPosted] = mutableSetOf(userId)
-                } else {
-                    usersWhoPostedByDate[dateUserPosted]!!.add(userId)
-                }
-            }
-        }
-
-        var date = startOfRelevantHistory
-        // doing this sequentially, instead of in parallel, so the persistence channel is easier to read
-        while (date <= yesterday) {
-            if (date !in persistedDates) {
-                val usersWhoPosted = usersWhoPostedByDate[date] ?: setOf()
-                logger.debug { "Persisting missing post history for $date" }
-                persistPostHistoryForDay(date, usersWhoPosted)
-            }
-            date = date.plusDays(1)
-        }
     }
 
     /** The earliest date we need to scan back to, given what we already have found in the persistence channel */
